crawler/requirements.txt
outscraper>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0
crawl4ai>=0.2.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
anthropic>=0.40.0
httpx>=0.27.0
